feat: set the ID and description for VIP 204
